Create a "core" and "full" version of the package
This commit follows a request for a lighter version of CleverCSV
that does not include as many dependencies (#10). With this commit
we create a "full" version by leveraging the extra_requires part
of the setup.py file, and leave the drop-in replacement of the csv
library as the "core" package. To make the core version of the
package lighter, I've also moved the command line interface
dependencies to the full version of the package.