Disable dialect pruning by default
This commit disables pruning the list of potential dialects
by default. This step can be quite expensive for large files,
and the added benefit is small, as the dialects that will
be pruned are generally few and aren't likely to be the best
performing ones anyway. By disabling this step by default
we (statistically) significantly improve the average detection
time.